Extracted some components to their own files. and imported them
